Change type of TC_uaf in daqSensors from 32-bit to 16-bit int
On 2019-10-18 15:52, David Knight wrote:
> We just tested the TC, the data memeber TC_uaf on line 14 of serial.h
needs to changed to an int16_t. After that change is made the TC works
great

+++ src/include/serial.h
@@ -11,7 +11,7 @@
 
 struct daqSensors {
 	uint32_t timestamp;
-	int32_t TC_uaf;
+	int16_t TC_uaf;
 	uint8_t HALL_methane, HALL_lox;
 	uint16_t BATT_voltage;
 	int16_t PT_methane, PT_lox, PT_helium, PT_chamber, PT_heliumReg;
